+++ colpali_engine/collators/visual_retriever_collator.py
@@ -1,8 +1,9 @@
-from typing import Any, Dict, List, Union, cast
+import random
+from typing import Any, Dict, List, Union
 
 from PIL.Image import Image
 
-from colpali_engine.models.idefics_2 import ColIdefics2Processor
+from colpali_engine.data.dataset import Document, IRDataset
 from colpali_engine.models.paligemma import ColPaliProcessor
 from colpali_engine.utils.processing_utils import BaseVisualRetrieverProcessor
 
@@ -19,6 +20,15 @@ class VisualRetrieverCollator:
     Collator for training vision retrieval models.
     """
 
+    # Input keys
+    query_key = IRDataset.QUERY_KEY
+    pos_target_key = IRDataset.POS_TARGET_KEY
+    neg_target_key = IRDataset.NEG_TARGET_KEY
+    # Prefixes
+    query_prefix = "query_"
+    pos_doc_prefix = "doc_"
+    neg_doc_prefix = "neg_doc_"
+
     def __init__(
         self,
         processor: BaseVisualRetrieverProcessor,
@@ -29,7 +39,7 @@ class VisualRetrieverCollator:
         self.image_token_id = None
 
         # If processor is one of the supported types, extract the <image> token id.
-        if isinstance(self.processor, (ColPaliProcessor, ColIdefics2Processor)):
+        if isinstance(self.processor, (ColPaliProcessor,)):
             image_token = "<image>"
             try:
                 idx = self.processor.tokenizer.additional_special_tokens.index(image_token)
@@ -43,44 +53,70 @@ class VisualRetrieverCollator:
             self.processor.tokenizer.padding_side = "right"
 
     def __call__(self, examples: List[Dict[str, Any]]) -> Dict[str, Any]:
-        texts_query: List[Union[str, None]] = []
-        images: List[Image] = []
-        neg_images: List[Image] = []
+        queries: List[Union[None, str, Image]] = []
+        pos_targets: List[Union[str, Image]] = []
+        neg_targets: List[Union[str, Image]] = []
 
         # Parse the examples.
         for example in examples:
-            query = example.get("query")
-            texts_query.append(query)
+            query = example.get(self.query_key)
+            sampled_query = random.choice(query) if isinstance(query, list) else query
+            queries.append(sampled_query)
 
-            image = example.get("image")
-            if image is None:
+            pos_tgt = example.get(self.pos_target_key)
+            if pos_tgt is not None:
+                sample_pos = random.choice(pos_tgt) if isinstance(pos_tgt, list) else pos_tgt
+                pos_targets.append(sample_pos)
+            else:
                 raise ValueError("Image is None - This collator does not support None images yet.")
-            images.append(cast(Image, image))
 
-            neg_image = example.get("neg_image")
-            if neg_image is not None:
-                neg_images.append(cast(Image, neg_image))
-
-        # Process images.
-        batch_doc = self.processor.process_images(images=images)
-        batch_neg_doc = self.processor.process_images(images=neg_images) if neg_images else None
+            neg_tgt = example.get(self.neg_target_key)
+            if neg_tgt is not None:
+                sampled_neg = random.choice(neg_tgt) if isinstance(neg_tgt, list) else neg_tgt
+                neg_targets.append(sampled_neg)
 
         # Process queries.
-        if all(q is None for q in texts_query):
+        if all(q is None for q in queries):
             batch_query = None
-        elif any(q is None for q in texts_query):
+        elif any(q is None for q in queries):
             raise ValueError("Some queries are None. This collator does not support None queries yet.")
         else:
-            batch_query = self.processor.process_queries(
-                queries=cast(List[str], texts_query),
-                max_length=self.max_length,
-            )
+            batch_query = self.auto_collate(queries, prefix=self.query_prefix)
 
+        # Process targets.
+        batch_pos_target = self.auto_collate(pos_targets, prefix=self.pos_doc_prefix)
+        batch_neg_target = self.auto_collate(neg_targets, prefix=self.neg_doc_prefix) if neg_targets else {}
+
+        return {
+            **batch_query,
+            **batch_pos_target,
+            **batch_neg_target,
+        }
+
+    def auto_collate(self, batch: List[Document], prefix: str = "") -> Dict[str, Any]:
+        """Automatically collate a batch of documents."""
+        # Convert Document objects to their underlying data.
+        batch = [b.item if isinstance(b, Document) else b for b in batch]
+        if isinstance(batch[0], str):
+            return self.collate_texts(batch, prefix=prefix)
+        elif isinstance(batch[0], Image):
+            return self.collate_images(batch, prefix=prefix)
+        else:
+            raise ValueError(f"Unsupported batch type: {type(batch[0])}. Expected str or Image.")
+
+    def collate_images(self, images: List[Image], prefix: str = "") -> Dict[str, Any]:
+        """Collate images into a batch."""
+        # Process images.
+        batch_im = self.processor.process_images(images=images)
         # Prefix keys to avoid collisions.
-        batch_all = prefix_keys(batch_doc, "doc_")
-        if batch_query:
-            batch_all.update(prefix_keys(batch_query, "query_"))
-        if batch_neg_doc:
-            batch_all.update(prefix_keys(batch_neg_doc, "neg_doc_"))
+        return prefix_keys(batch_im, prefix)
 
-        return batch_all
+    def collate_texts(self, texts: List[str], prefix: str = "") -> Dict[str, Any]:
+        """Collate texts into a batch."""
+        # Process texts.
+        batch_text = self.processor.process_queries(
+            queries=texts,
+            max_length=self.max_length,
+        )
+        # Prefix keys to avoid collisions.
+        return prefix_keys(batch_text, prefix)