+++ colpali_engine/utils/dataset_transformation.py
@@ -1,7 +1,10 @@
 import os
-from typing import List, Tuple, cast
+from typing import List, Literal, Tuple, cast
 
 from datasets import Dataset, DatasetDict, concatenate_datasets, load_dataset
+
+from colpali_engine.data.corpus import LocalCorpus, SimpleCorpus
+from colpali_engine.data.dataset import IRColumn, IRDataset
 
 USE_LOCAL_DATASET = os.environ.get("USE_LOCAL_DATASET", "1") == "1"
 
@@ -19,6 +22,83 @@ def load_train_set() -> DatasetDict:
     base_path = "./data_dir/" if USE_LOCAL_DATASET else "vidore/"
     ds_dict = cast(DatasetDict, load_dataset(base_path + ds_path))
     return ds_dict
+
+
+def load_train_set_ir() -> IRDataset:
+    """Returns the query dataset, then the anchor dataset with the documents, then the dataset type"""
+    corpus = SimpleCorpus(
+        corpus_data=load_dataset("manu/colpali-corpus", split="train"),
+        id_column=None,
+    )
+
+    dataset = load_dataset("manu/colpali-queries", split="train")
+
+    print("Dataset size:", len(dataset))
+    # filter out queries with "gold_in_top_100" == False
+    dataset = dataset.filter(lambda x: x["gold_in_top_100"], num_proc=16)
+    # keep only top 50 negative passages
+    dataset = dataset.map(lambda x: {"negative_passages": x["negative_passages"][:50]})
+    print("Dataset size after filtering:", len(dataset))
+
+    train_dataset = IRDataset(
+        data=dataset,
+        query_column="query",
+        pos_target_column=IRColumn("positive_passages", corpus_column="image"),
+        corpus=corpus,
+    )
+
+    return train_dataset
+
+
+def load_train_set_ir_negs() -> IRDataset:
+    """Returns the query dataset, then the anchor dataset with the documents, then the dataset type"""
+    corpus = SimpleCorpus(
+        corpus_data=load_dataset("manu/colpali-corpus", split="train"),
+        id_column=None,
+    )
+
+    dataset = load_dataset("manu/colpali-queries", split="train")
+
+    print("Dataset size:", len(dataset))
+    # filter out queries with "gold_in_top_100" == False
+    dataset = dataset.filter(lambda x: x["gold_in_top_100"], num_proc=16)
+    # keep only top 50 negative passages
+    dataset = dataset.map(lambda x: {"negative_passages": x["negative_passages"][:50]})
+    print("Dataset size after filtering:", len(dataset))
+
+    train_dataset = IRDataset(
+        data=dataset.select(range(500, len(dataset))),
+        query_column="query",
+        pos_target_column=IRColumn("positive_passages", corpus_column="image"),
+        neg_target_column=IRColumn("negative_passages", corpus_column="image"),
+        corpus=corpus,
+    )
+
+    return train_dataset
+
+
+def load_mmeb(subset: str, corpus_path: str, type: Literal["t2i", "i2t"], use_negatives: bool = False) -> IRDataset:
+    corpus = LocalCorpus(corpus_path=corpus_path)
+    dataset = load_dataset("TIGER-Lab/MMEB-train", subset, split="original")
+
+    if type == "t2i":
+        query_col = "qry"
+        pos_col = IRColumn("pos_image_path", corpus_column="doc")
+        neg_col = IRColumn("neg_image_path", corpus_column="doc") if use_negatives else None
+    elif type == "i2t":
+        query_col = IRColumn("qry_image_path", corpus_column="doc")
+        pos_col = "pos_text"
+        neg_col = "neg_text" if use_negatives else None
+    else:
+        raise ValueError("type must be either 't2i' or 'i2t'")
+
+    return IRDataset(
+        data=dataset,
+        query_column=query_col,
+        pos_target_column=pos_col,
+        neg_target_column=neg_col,
+        corpus=corpus,
+    )
 
 
 def load_train_set_detailed() -> DatasetDict:
@@ -111,27 +191,6 @@ def load_wikiss() -> Tuple[DatasetDict, Dataset, str]:
     anchor_ds = cast(Dataset, load_dataset(base_path + "wiki-ss-corpus", split="train"))
 
     return ds_dict, anchor_ds, "wikiss"
-
-
-def load_train_set_ir_negs() -> Tuple[DatasetDict, Dataset, str]:
-    """Returns the query dataset, then the anchor dataset with the documents, then the dataset type"""
-    base_path = "./data_dir/" if USE_LOCAL_DATASET else "manu/"
-    dataset = cast(Dataset, load_dataset(base_path + "colpali-queries", split="train"))
-
-    print("Dataset size:", len(dataset))
-    # filter out queries with "gold_in_top_100" == False
-    dataset = dataset.filter(lambda x: x["gold_in_top_100"], num_proc=16)
-    print("Dataset size after filtering:", len(dataset))
-
-    # keep only top 20 negative passages
-    dataset = dataset.map(lambda x: {"negative_passages": x["negative_passages"][:20]})
-
-    dataset_eval = dataset.select(range(500))
-    dataset = dataset.select(range(500, len(dataset)))
-    ds_dict = DatasetDict({"train": dataset, "test": dataset_eval})
-
-    anchor_ds = cast(Dataset, load_dataset(base_path + "colpali-corpus", split="train"))
-    return ds_dict, anchor_ds, "vidore"
 
 
 def load_train_set_with_docmatix() -> DatasetDict: