+++ scripts/train/train_colbert.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 
 import configue
@@ -6,23 +7,36 @@ import typer
 from colpali_engine.trainer.colmodel_training import ColModelTraining, ColModelTrainingConfig
 from colpali_engine.utils.gpu_stats import print_gpu_utilization
 
+app = typer.Typer(pretty_exceptions_enable=False)
 
+
+@app.command()
 def main(config_file: Path) -> None:
+    """
+    Training script for ColVision models.
+
+    Args:
+        config_file (Path): Path to the configuration file.
+    """
     print_gpu_utilization()
+
     print("Loading config")
     config = configue.load(config_file, sub_path="config")
+
     print("Creating Setup")
     if isinstance(config, ColModelTrainingConfig):
-        app = ColModelTraining(config)
+        training_app = ColModelTraining(config)
     else:
         raise ValueError("Config must be of type ColModelTrainingConfig")
 
     if config.run_train:
         print("Training model")
-        app.train()
-        app.save(config_file=config_file)
+        training_app.train()
+        training_app.save()
+        os.system(f"cp {config_file} {training_app.config.output_dir}/training_config.yml")
+
     print("Done!")
 
 
 if __name__ == "__main__":
-    typer.run(main)
+    app()