+++ CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [0.3.10] - 2025-04-18
+
+### Added
+
+- Add `LambdaTokenPooler` to allow for custom token pooling functions.
+- Added training losses with negatives to InfoNCE type losses
+
 ### Changed
 
-- [Breaking] Make `pool_factor` an argument of `pool_embeddings` instead of a `HierarchicalTokenPooler` class attribute
+- Fix similarity map helpers for ColQwen2 and ColQwen2.5.
+- [Breaking] (minor) Remove support for Idefics2-based models.
+- Disable multithreading in `HierarchicalTokenPooler` if `num_workers` is not provided or is 1.
+- [Breaking] (minor) Make `pool_factor` an argument of `pool_embeddings` instead of a `HierarchicalTokenPooler` class attribute
+- Bump dependencies for transformers, torch, peft, pillow, accelerate, etc...
 
 ## [0.3.9] - 2025-04-03